Application - ajout home page + navigation entre les deux premiers fragments
