+++ Application/app/src/main/java/fr/mastersid/deliens/home/ui/homePage/HomePageFragment.kt
@@ -1,0 +1,34 @@
+package fr.mastersid.deliens.home.ui.homePage
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.navigation.fragment.findNavController
+import dagger.hilt.android.AndroidEntryPoint
+import fr.mastersid.deliens.home.databinding.FragmentHomePageBinding
+
+@AndroidEntryPoint
+class HomePageFragment : Fragment() {
+
+    private lateinit var binding: FragmentHomePageBinding
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        binding = FragmentHomePageBinding.inflate(inflater)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.goToListEstimation.setOnClickListener() {
+            val action = HomePageFragmentDirections.actionHomePageFragmentToListEstimation()
+            findNavController().navigate(action)
+        }
+    }
+}
